fix: Resolve build errors from previous commits
- SettingsViewModel: Fix Action<float> to Action<double> type mismatch
- ProfileViewModel: Update using alias to PresenceActivityType
- ServerAdminViewModel: Remove duplicate Cleanup method
- AuthController: Fix IsValidEmail method placement (was outside class)
- ScreenStreamController: Change FrameCount/TotalBytes to fields for Interlocked

+++ src/VeaMarketplace.Client/ViewModels/SettingsViewModel.cs
@@ -13,7 +13,7 @@ public partial class SettingsViewModel : BaseViewModel
     private readonly IVoiceService _voiceService;
 
     // Store event handler for proper unsubscription
-    private readonly Action<float> _onLocalAudioLevel;
+    private readonly Action<double> _onLocalAudioLevel;
 
     [ObservableProperty]
     private ObservableCollection<AudioDeviceItem> _inputDevices = new();
